requirements.txt
numpy
Pillow
tabulate
